feat: structure of cargo near `build` command with interactive subcommand
